src/test/java/com/codingquestions/StringII/LongestSubarrayContainsOnly1sTest.java
package com.codingquestions.StringII;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;

import java.util.Arrays;
import java.util.List;

import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Tag;
import org.junit.jupiter.api.Test;

// import the class that you want to test here
import com.codingquestions.app.StringII.ThreeSlidingWIndowQuestions.LongestSubarrayContainsOnly1s;

public class LongestSubarrayContainsOnly1sTest {

    @BeforeAll
    static void setup() {
    }

    @BeforeEach
    void setupThis() {
    }

    @Test
    void testOne() {
        LongestSubarrayContainsOnly1s instance = new LongestSubarrayContainsOnly1s();

        assertEquals(2, instance.longestConsecutiveOnesMethod2(new int[] { 1, 0, 0, 1, 1 }, 0));
    }

    @AfterEach
    void tearThis() {
    }

    @AfterAll
    static void tear() {
    }
}